09._Environment_Variables/02._python/pyproject.toml
[tool.poetry]
name = "02-python"
version = "0.1.0"
description = ""
authors = ["Marcus <marc303i@stud.kea.dk>"]
readme = "README.md"
packages = [{include = "02"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
